+++ VPN-Sentinel@claudiux/files/VPN-Sentinel@claudiux/applet.js
@@ -1038,7 +1038,8 @@ class VPNSentinel extends Applet.TextIconApplet {
 
     let id = setTimeout( () => {
       clearTimeout(id);
-      Util.spawnCommandLine("cinnamon-settings applets %s -t %s %s".format(UUID, ""+tabFlags, ""+instanceId));
+      //~ Util.spawnCommandLine("cinnamon-settings applets %s -t %s %s".format(UUID, ""+tabFlags, ""+instanceId));
+      Util.spawnCommandLine(`xlet-settings -t ${tabFlags} -i ${instanceId} applet ${UUID}`);
     }, 500); // 500 ms
   } // End of sort_flaglist
 
@@ -1157,14 +1158,16 @@ class VPNSentinel extends Applet.TextIconApplet {
     // button VPN Policy...
     let configure1 = new PopupMenu.PopupIconMenuItem(_("VPN Policy") + "...", "system-run", St.IconType.SYMBOLIC);
     configure1.connect("activate", () => {
-      Util.spawnCommandLine("cinnamon-settings applets %s -t %s %s".format(UUID, ""+this.tabNumberOfVPNPolicy, ""+this.instanceId));
+      //~ Util.spawnCommandLine("cinnamon-settings applets %s -t %s %s".format(UUID, ""+this.tabNumberOfVPNPolicy, ""+this.instanceId));
+      Util.spawnCommandLine(`xlet-settings -t ${this.tabNumberOfVPNPolicy} -i ${this.instanceId} applet ${UUID}`);
     });
     this.menu.addMenuItem(configure1);
 
     // button Internet Apps Manager...
     let configure2 = new PopupMenu.PopupIconMenuItem(_("Internet Apps Manager") + "...", "system-run", St.IconType.SYMBOLIC);
     configure2.connect("activate", () => {
-      Util.spawnCommandLine("cinnamon-settings applets %s -t %s %s".format(UUID, ""+this.tabNumberOfInternetPolicy, ""+this.instanceId));
+      //~ Util.spawnCommandLine("cinnamon-settings applets %s -t %s %s".format(UUID, ""+this.tabNumberOfInternetPolicy, ""+this.instanceId));
+      Util.spawnCommandLine(`xlet-settings -t ${this.tabNumberOfInternetPolicy} -i ${this.instanceId} applet ${UUID}`);
     });
     this.menu.addMenuItem(configure2);